% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_points.R
\docType{data}
\name{sample_points}
\alias{sample_points}
\title{An example set of five 2D points}
\format{
A matrix with 5 rows and 2 columns:
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\usage{
data("sample_points")
}
\description{
A 5 x 2 matrix of (x, y) coordinates to demonstrate the function.
}
\keyword{internal}
